+++ Lab4/DistributedFileSystem/DistributedFileSystem.Server/Services/ValueObjects/Node.cs
@@ -4,7 +4,7 @@ namespace DistributedFileSystem.Services.ValueObjects;
 
 public class Node
 {
-    public Node(string name, IPAddress ipAddress, int port, int size)
+    public Node(string name, IPAddress ipAddress, int port, long size)
     {
         Name = name;
         Size = size;
@@ -14,8 +14,8 @@ public class Node
     }
 
     public string Name { get; }
-    public int Size { get; }
-    public int FreeMemory { get; set; }
+    public long Size { get; }
+    public long FreeMemory { get; set; }
     public IPAddress IpAddress { get; }
     public int Port { get; }
 }
